feat: Implement logout functionality changes includes: - Added a new route for logout - Implemented a middleware (isLoggedIn) - moved `isAlreadyLoggedIn` middleware from auth.controller.ts
